Run importability annotation on every content channel import page load, if needed.

+++ kolibri/plugins/device_management/assets/src/modules/wizard/utils.js
@@ -19,39 +19,66 @@ export function getRemoteChannelBundleByToken(token) {
 }
 
 /**
- * Starts Task that downloads a Channel Metadata database.
+ * Starts Tasks that either downloads a Channel Metadata database,
+ * or annotate the content importability for the current import method
  * NOTE: cannot be normally dispatched as an action, since it uses
  * waitForTaskToComplete (which relies on the store singleton with a .watch method)
  *
  */
-export function downloadChannelMetadata(store) {
+export function readyChannelMetadata(store, download = true) {
   const { transferredChannel, selectedDrive, selectedPeer } = store.state.manageContent.wizard;
+  let skipTaskCheck;
   let promise;
-  if (store.getters['manageContent/wizard/inLocalImportMode']) {
-    promise = TaskResource.startDiskChannelImport({
-      channel_id: transferredChannel.id,
-      drive_id: selectedDrive.id,
-    });
-  } else if (store.getters['manageContent/wizard/inRemoteImportMode']) {
-    promise = TaskResource.startRemoteChannelImport({
-      channel_id: transferredChannel.id,
-    });
-  } else if (store.getters['manageContent/wizard/inPeerImportMode']) {
-    promise = TaskResource.startRemoteChannelImport({
-      channel_id: transferredChannel.id,
-      baseurl: selectedPeer.base_url,
-    });
+  if (download) {
+    if (store.getters['manageContent/wizard/inLocalImportMode']) {
+      promise = TaskResource.startDiskChannelImport({
+        channel_id: transferredChannel.id,
+        drive_id: selectedDrive.id,
+      });
+    } else if (store.getters['manageContent/wizard/inRemoteImportMode']) {
+      promise = TaskResource.startRemoteChannelImport({
+        channel_id: transferredChannel.id,
+      });
+    } else if (store.getters['manageContent/wizard/inPeerImportMode']) {
+      promise = TaskResource.startRemoteChannelImport({
+        channel_id: transferredChannel.id,
+        baseurl: selectedPeer.base_url,
+      });
+    } else {
+      return Error('Channel Metadata is only downloaded when importing');
+    }
   } else {
-    return Error('Channel Metadata is only downloaded when importing');
+    if (store.getters['manageContent/wizard/inLocalImportMode']) {
+      promise = TaskResource.startAnnotateImportability({
+        channel_id: transferredChannel.id,
+        drive_id: selectedDrive.id,
+      });
+    } else if (store.getters['manageContent/wizard/inRemoteImportMode']) {
+      promise = Promise.resolve();
+      skipTaskCheck = true;
+    } else if (store.getters['manageContent/wizard/inPeerImportMode']) {
+      promise = TaskResource.startAnnotateImportability({
+        channel_id: transferredChannel.id,
+        baseurl: selectedPeer.base_url,
+      });
+    } else {
+      return Error('Channel Metadata is only downloaded when importing');
+    }
   }
   promise = promise.catch(() => Promise.reject({ errorType: ErrorTypes.CONTENT_DB_LOADING_ERROR }));
 
   return promise
     .then(task => {
+      if (skipTaskCheck) {
+        return Promise.resolve();
+      }
       // NOTE: store.watch is not available to dispatched actions
       return waitForTaskToComplete(store, task.entity.id);
     })
     .then(completedTask => {
+      if (skipTaskCheck) {
+        return getChannelWithContentSizes(transferredChannel.id);
+      }
       const { taskId, cancelled } = completedTask;
       if (taskId && !cancelled) {
         return TaskResource.deleteFinishedTasks().then(() => {
